Move display and position to its own CSS block
The issue was that prior to this, `display: inline-block` and `position:
relative` was only being applied to when .navigation-item was in a
:hover state. Aside from that, `.navigation-item` had a property of
`display: flex` that it was inheriting from `.navigation`.

Once you move those two properties to its own CSS block, the transition
works as intended. (Btw, it's a nice little effect)